Crate fuse

Re-exports

pub use mount::mount;

Modules

filesystem

mount

sys

worker

Structs

Server

Enums

Error
Errors that may occur during the creation or operation of an Fs device.

Traits

Mapper
A trait for memory mapping for DAX.

Reader
A trait for reading from the underlying FUSE endpoint.

Writer
A trait for writing to the underlying FUSE endpoint. The FUSE device expects the write operation to happen in one write transaction. Since there are cases when data needs to be generated earlier than the header, it implies the writer implementation to keep an internal buffer. The buffer then can be flushed once header and data are both prepared.

Type Definitions

Result