Module fuse::filesystem

Re-exports

pub use crate::sys::FsOptions;

pub use crate::sys::IoctlFlags;

pub use crate::sys::IoctlIovec;

pub use crate::sys::OpenOptions;

pub use crate::sys::RemoveMappingOne;

pub use crate::sys::SetattrValid;

pub use crate::sys::ROOT_ID;

Structs

Context
Additional context associated with requests.

DirEntry
Represents information about an entry in a directory.

Entry
Information about a path in the filesystem.

Enums

GetxattrReply
A reply to a getxattr method call.

IoctlReply
A reply to an ioctl method call.

ListxattrReply
A reply to a listxattr method call.

Traits

DirectoryIterator
A trait for iterating over the contents of a directory. This trait is needed because rust doesn't support generic associated types, which means that it's not possible to implement a regular iterator that yields a DirEntry due to its generic lifetime parameter.

FileSystem
The main trait that connects a file system with a transport.

ZeroCopyReader
A trait for directly copying data from the fuse transport into a File without first storing it in an intermediate buffer.

ZeroCopyWriter
A trait for directly copying data from a File into the fuse transport without first storing it in an intermediate buffer.