Trait fuse::filesystem::DirectoryIterator

pub trait DirectoryIterator {
    fn next(&mut self) -> Option<DirEntry<'_>>;
}

A trait for iterating over the contents of a directory. This trait is needed because rust doesn’t support generic associated types, which means that it’s not possible to implement a regular iterator that yields a DirEntry due to its generic lifetime parameter.

Required methods

fn next(&mut self) -> Option<DirEntry<'_>>

Returns the next entry in the directory or None if there are no more.