Module fuse::server

Structs

Server

SplitInclusive

Constants

DIRENT_PADDING

Traits

Mapper

A trait for memory mapping for DAX.

Reader

A trait for reading from the underlying FUSE endpoint.

Writer

A trait for writing to the underlying FUSE endpoint. The FUSE device expects the write operation to happen in one write transaction. Since there are cases when data needs to be generated earlier than the header, it implies the writer implementation to keep an internal buffer. The buffer then can be flushed once header and data are both prepared.

Functions

add_dirent

bytes_to_cstr

finish_ioctl

reply_error

reply_ok

reply_readdir

retry_ioctl

split_inclusive