Struct fuse::BackgroundSession

pub struct BackgroundSession<'a> {
    pub mountpoint: PathBuf,
    pub guard: JoinGuard<'a, ()>,
}

The background session data structure

Fields

mountpoint

Path of the mounted filesystem

guard

Thread guard of the background session

Methods

impl<'a> BackgroundSession<'a>

unsafe fn new<FS: Filesystem + Send + 'a>(se: Session<FS>) -> Result<BackgroundSession<'a>>

Create a new background session for the given session by running its session loop in a background thread. If the returned handle is dropped, the filesystem is unmounted and the given session ends.