Struct fuse::ReplyDirectory

pub struct ReplyDirectory {
    // some fields omitted
}

Directory reply

Methods

impl ReplyDirectory

fn new<S: ReplySender>(unique: u64, sender: S, size: usize) -> ReplyDirectory

Creates a new ReplyDirectory with a specified buffer size.

fn add<T: AsRef<OsStr>>(&mut self, ino: u64, offset: u64, kind: FileType, name: T) -> bool

Add an entry to the directory reply buffer. Returns true if the buffer is full. A transparent offset value can be provided for each entry. The kernel uses these value to request the next entries in further readdir calls

fn ok(self)

Reply to a request with the filled directory buffer

fn error(self, err: c_int)

Reply to a request with the given error code