Struct fuse::ReplyStatfs

pub struct ReplyStatfs {
    // some fields omitted
}

Statfs Reply

Methods

impl ReplyStatfs

fn statfs(self, blocks: u64, bfree: u64, bavail: u64, files: u64, ffree: u64, bsize: u32, namelen: u32, frsize: u32)

Reply to a request with the given open result

fn error(self, err: c_int)

Reply to a request with the given error code