Crate fuse

FUSE userspace library implementation (as of libosxfuse-2.5.5).

This is an improved rewrite of the FUSE userspace library (lowlevel interface) to fully take advantage of Rust's architecture. The only thing we rely on in the real libfuse are mount and unmount calls which are needed to establish a fd to talk to the kernel driver.

Modules

consts

Structs

BackgroundSession
The background session data structure

FileAttr
File attributes

ReplyAttr

ReplyBmap

ReplyCreate

ReplyData

ReplyDirectory

ReplyEmpty

ReplyEntry

ReplyLock

ReplyOpen

ReplyStatfs

ReplyWrite

ReplyXTimes

Request
Request data structure

Session
The session data structure

Enums

FileType
File types

Constants

FUSE_ROOT_ID

Traits

Filesystem
Filesystem trait.

Reply
Generic reply trait

Functions

mount
Mount the given filesystem to the given mountpoint. This function will not return until the filesystem is unmounted.

spawn_mount
Mount the given filesystem to the given mountpoint. This function spawns a background thread to handle filesystem operations while being mounted and therefore returns immediately. The returned handle should be stored to reference the mounted filesystem. If it's dropped, the filesystem will be unmounted.