Module fuse::sys

Structs

AccessIn

Attr

AttrOut

BatchForgetIn

BmapIn

BmapOut

ChromeOsTmpfileIn

CopyFileRangeIn

CreateIn

Dirent

Direntplus

EntryOut

FallocateIn

FileLock

FlushIn

ForgetIn

ForgetOne

FsOptions

A bitfield passed in as a parameter to and returned from the init method of the FileSystem trait.

FsyncIn

GetattrIn

GetxattrIn

GetxattrOut

InHeader

InitIn

InitOut

InterruptIn

IoctlFlags

IoctlIn

IoctlIovec

Describes a region of memory in the address space of the process that made the ioctl syscall. Similar to libc::iovec but uses u64s for the address and the length.

IoctlOut

Kstatfs

LinkIn

LkIn

LkOut

LseekIn

LseekOut

MkdirIn

MknodIn

NotifyDeleteOut

NotifyInvalEntryOut

NotifyInvalInodeOut

NotifyPollWakeupOut

NotifyRetrieveIn

NotifyStoreOut

Notify_Retrieve_Out

OpenIn

OpenOptions

Options controlling the behavior of files opened by the server in response to an open or create request.

OpenOut

OutHeader

PollIn

PollOut

ReadIn

ReleaseIn

RemoveMappingIn

RemoveMappingOne

Rename2In

RenameIn

SetUpMappingFlags

SetUpMappingIn

SetattrIn

SetattrValid

SetxattrIn

StatfsOut

WriteIn

WriteOut

Enums

NotifyOpcode

Opcode

Constants

FATTR_FH

FATTR_LOCKOWNER

FUSE_COMPAT_22_INIT_OUT_SIZE

FUSE_COMPAT_ATTR_OUT_SIZE

FUSE_COMPAT_ENTRY_OUT_SIZE

FUSE_COMPAT_INIT_OUT_SIZE

FUSE_COMPAT_MKNOD_IN_SIZE

FUSE_COMPAT_STATFS_SIZE

FUSE_COMPAT_WRITE_IN_SIZE

FUSE_MIN_READ_BUFFER

The read buffer is required to be at least 8k, but may be much larger.

GETATTR_FH

IOCTL_MAX_IOV

Maximum of in_iovecs + out_iovecs

KERNEL_MINOR_VERSION

Minor version number of this interface.

KERNEL_VERSION

Version number of this interface.

LK_FLOCK

OLDEST_SUPPORTED_KERNEL_MINOR_VERSION

Oldest supported minor version of the fuse interface.

POLL_SCHEDULE_NOTIFY

Request poll notify.

READ_LOCKOWNER

RELEASE_FLOCK_UNLOCK

RELEASE_FLUSH

ROOT_ID

The ID of the inode corresponding to the root directory of the file system.

WRITE_CACHE

Delayed write from page cache, file handle is guessed.

WRITE_KILL_PRIV

Kill the suid and sgid bits.

WRITE_LOCKOWNER

lock_owner field is valid.